// web/src/api/generated/user.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type UserRes = {
  id: bigint;
  first_name: string;
  last_name: string;
  email: string;
  created_at: string;
  updated_at: string;
};

export type UserUpdateReq = { first_name: string; last_name: string };

export type UsersRes = { data: Array<UserRes> };